pyproject.toml
[tool.poetry]
name = "cs-toolkit"
version = "0.1.0"
description = "(Roughly) Minimalist framework for exploring code switching data"
authors = ["ZechyW <ZechyW@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
djangorestframework = "^3.9"
channels = "^2.1"
channels-redis = "^2.3"
whitenoise = "^4.1"
django-model-utils = "^3.1"
sphinx = "^1.8"
sphinx-js = "^2.7"
sphinx_rtd_theme = "^0.4.2"
m2r = "^0.2.1"
djangorestframework-queryfields = "^1.0"
bootstrap-admin = "0.4.0"
djangorestframework-datatables = "^0.4.1"
requests = "^2.21"
autobahn = "^19.1"
twisted = "^18.9"
dramatiq = {version = "^1.5",extras = ["watch", "redis"]}
django-dramatiq = "^0.5.3"
django-mptt = "^0.9.1"
psycopg2-binary = "^2.7"
djangorestframework-recursive = "^0.1.2"
backoff = "^1.8"
pywin32 = {version = "^227",optional = true}

[tool.poetry.dev-dependencies]
pylint = "^2.2"
white = "^0.1.2"
ipython = {version = "^7.3",extras = ["shell"]}
black = "^19.10b0"

[tool.poetry.extras]
windows = ["pywin32"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
